combined preparation of Slater determinants and fermionic Gaussian states into one file
